Move cas1 migration code into cas1 package
